ml_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
langchain
langchain-core
langchain-groq
python-dotenv